♻️ M9JY3M task: decompose run command
