+++ .agentplane/tasks/202605282136-M9JY3M/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605282136-M9JY3M`
+Title: Task run command decomposition
+Canonical task record: `.agentplane/tasks/202605282136-M9JY3M/README.md`
+
+## Summary
+
+Task run command decomposition
+
+Decompose packages/agentplane/src/commands/task/run.command.ts into focused task-run CLI helper modules without changing runner lifecycle, status, inspect, logs, wait, or JSON/text output behavior. Reduce the command file below the hotspot warning threshold when safe.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/task/run.command.ts into focused task-run CLI helper modules without changing runner lifecycle, status, inspect, logs, wait, or JSON/text output behavior. Reduce the command file below the hotspot warning threshold when safe.
+- Out of scope: unrelated refactors not required for "Task run command decomposition".
+
+## Verification
+
+- State: ok
+- Note: Task run command decomposition verified locally.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T21:36:56.557Z
+- Branch: task/202605282136-M9JY3M/task-run-command-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
