+++ .agentplane/tasks/202605282136-M9JY3M/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 M9JY3M task: Task run command decomposition [202605282136-M9JY3M]
