+++ .agentplane/tasks/202605282136-M9JY3M/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T21:36:56.557Z
+
+## Task
+
+- Task: `202605282136-M9JY3M`
+- Title: Task run command decomposition
+- Status: DOING
+- Branch: `task/202605282136-M9JY3M/task-run-command-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282136-M9JY3M/README.md`
+
+## Verification
+
+- State: ok
+- Note: Task run command decomposition verified locally.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T21:36:56.557Z
+- Branch: task/202605282136-M9JY3M/task-run-command-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
